# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EduFlow API",
        "description": "Documentation API complète pour la plateforme LMS EduFlow. Gestion des cours, étudiants, enseignants, favoris, inscriptions et groupes.",
        "contact": {
            "name": "EduFlow Support",
            "email": "support@eduflow.dev"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "{protocol}://{host}:{port}/api",
            "description": "EduFlow API Server",
            "variables": {
                "protocol": {
                    "enum": [
                        "http",
                        "https"
                    ],
                    "default": "http"
                },
                "host": {
                    "enum": [
                        "localhost",
                        "api.eduflow.dev"
                    ],
                    "default": "localhost"
                },
                "port": {
                    "enum": [
                        "8000",
                        "443"
                    ],
                    "default": "8000"
                }
            }
        }
    ],
    "paths": {
...
        "/courses": {
            "get": {
                "tags": [
                    "Courses"
                ],
                "summary": "Lister tous les cours",
                "description": "Récupère la liste de tous les cours disponibles. Les teachers voient uniquement leurs propres cours.",
                "operationId": "getCoursesList",
                "responses": {
                    "200": {
                        "description": "Liste des cours récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "courses": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "format": "int64",
                                                        "example": 1
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "PHP Avancé"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "Cours approfondi de PHP avec patterns modernes"
                                                    },
                                                    "price": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 49.99
                                                    },
                                                    "teacher_id": {
                                                        "type": "integer",
                                                        "format": "int64",
                                                        "example": 2
                                                    },
                                                    "category_id": {
                                                        "type": "integer",
                                                        "format": "int64",
                                                        "example": 1
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Courses"
                ],
                "summary": "Créer un nouveau cours",
                "description": "Crée un nouveau cours. Seuls les teachers peuvent créer des cours.",
                "operationId": "createCourse",
                "requestBody": {
                    "description": "Données du cours",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "price",
                                    "category_id"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "PHP Avancé"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Cours approfondi de PHP avec patterns modernes"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "minimum": 0,
                                        "example": 49.99
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "format": "int64",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Cours créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "format": "int64",
                                            "example": 1
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "PHP Avancé"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Cours approfondi de PHP avec patterns modernes"
                                        },
                                        "price": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 49.99
                                        },
                                        "teacher_id": {
                                            "type": "integer",
                                            "format": "int64",
                                            "example": 2
                                        },
                                        "category_id": {
                                            "type": "integer",
                                            "format": "int64",
                                            "example": 1
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation failed"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Accès refusé - Rôle teacher requis"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Bearer Token pour l'authentification JWT",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}